feat: modulize SubVerso (#168)
Adapts SubVerso to the module system.

Key aspects:
 1. Code for older Lean is generated with demodulize.py
 2. Tests are run with the appropriate version during CI
 3. On every commit to main, the no-modules branch is updated with the
    module-free version. The file .source-commit is committed with the
    original SHA from main.

One of the tests wouldn't run for the three oldest versions of Lean that are supported. It seems to be for reasons of runner/project config rather than because of the code being wrong, but it's of course impossible to know for sure. Rather than spend a ton of time, it was just disabled - it passed locally but not in CI, and doesn't seem worth it given that everything else passes.